styling fixes, changes to popup info
